moving definitions on store side
